Remove ApplicantBase
