+++ Mobile/Views/ViewApplicant.cs
@@ -15,15 +15,125 @@ using SM = Carmen.ShowModel;
 
 namespace Carmen.Mobile.Views
 {
-    internal class ViewApplicant : ApplicantBase
+    internal class ViewApplicant : ContentPage
     {
-        public ViewApplicant(ConnectionDetails show, int id, string first, string last)
-            : base(show, id, first, last)
-        { }
+        protected readonly ApplicantModel model;
+        protected readonly ConnectionDetails show;
+        protected ShowContext? context;
 
-        protected override View GenerateMainView()
+        public ViewApplicant(ConnectionDetails show, int id, string first, string last)
         {
-            //TODO (EDIT) move ApplicantBase code into ViewApplicant & remove EditApplicant/ApplicantBase
+            model = new(id);
+            this.show = show;
+            BindingContext = model;
+            Title = $"{first} {last}";
+            Loaded += ViewApplicant_Loaded;
+            Unloaded += ViewApplicant_Unloaded;
+
+            var loading = new ActivityIndicator { IsRunning = true };
+            loading.SetBinding(ActivityIndicator.IsVisibleProperty, new Binding(nameof(ApplicantModel.IsLoading)));
+
+            var main = new ScrollView
+            {
+                Content = GenerateMainView()
+            };
+
+            var grid = new Grid
+            {
+                Margin = 5,
+                RowSpacing = 5,
+                ColumnSpacing = 5,
+                RowDefinitions =
+                {
+                    new RowDefinition(GridLength.Star),
+                    new RowDefinition(GridLength.Auto)
+                },
+                ColumnDefinitions =
+                {
+                    new(GridLength.Star)
+                }
+            };
+            grid.Add(loading);
+            grid.Add(main);
+            var c = 0;
+            var back = new Button
+            {
+                Text = "Back",
+                BackgroundColor = Colors.Gray
+            };
+            back.Clicked += Back_Clicked;
+            grid.Add(back, row: 1, column: c++);
+            var delete = new Button
+            {
+                Text = "Delete",
+                BackgroundColor = Colors.Red
+            };
+            delete.Clicked += Delete_Clicked; ;
+            grid.ColumnDefinitions.Add(new(GridLength.Star));
+            grid.Add(delete, row: 1, column: c++);
+            var save = new Button
+            {
+                Text = "Save",
+            };
+            save.Clicked += Save_Clicked;
+            grid.ColumnDefinitions.Add(new(GridLength.Star));
+            grid.Add(save, row: 1, column: c++);
+            grid.SetColumnSpan(loading, c);
+            grid.SetColumnSpan(main, c);
+            Content = grid;
+        }
+
+        private async void ViewApplicant_Loaded(object? sender, EventArgs e)
+        {
+            context = ShowContext.Open(show);
+            var applicant = await Task.Run(() => context.Applicants.Single(a => a.ApplicantId == model.ApplicantId));
+            model.Loaded(applicant);
+            var image = await Task.Run(() => applicant.Photo); //TODO cache photos
+            var source = image == null ? null : await MauiImageSource(image);
+            model.LoadedPhoto(source);
+        }
+
+        static async Task<ImageSource?> MauiImageSource(SM.Image photo)
+        {
+            try
+            {
+                return await Task.Run(() =>
+                {
+                    Stream? stream = null;
+                    return ImageSource.FromStream(() =>
+                    {
+                        stream?.Dispose();
+                        stream = new MemoryStream(photo.ImageData);
+                        return stream;
+                    });
+                });
+            }
+            catch (Exception ex)
+            {
+                Log.Error(ex, $"Invalid image data of length {photo.ImageData.Length}, id {photo.ImageId}, {photo.Name}");
+                return null;
+            }
+        }
+
+        private void ViewApplicant_Unloaded(object? sender, EventArgs e)
+        {
+            context?.Dispose();
+            context = null;
+        }
+
+        private async void Back_Clicked(object? sender, EventArgs e)
+        {
+            if (context?.ChangeTracker.HasChanges() == true)
+            {
+                if (!await DisplayAlert($"Are you sure you want to discard your changes?", "", "Yes", "No"))
+                    return;
+            }
+            await Navigation.PopAsync();
+        }
+
+        private View GenerateMainView()
+        {
+            //TODO (EDIT) ViewApplicant to ApplicantDetails
             var fields = new ListView
             {
                 ItemTemplate = new DataTemplate(GenerateFieldDataTemplate),
@@ -74,30 +184,13 @@ namespace Carmen.Mobile.Views
             return layout;
         }
 
-        protected override IEnumerable<View> GenerateExtraButtons()
-        {
-            var delete = new Button
-            {
-                Text = "Delete",
-                BackgroundColor = Colors.Red
-            };
-            delete.Clicked += Delete_Clicked; ;
-            yield return delete;
-            var save = new Button
-            {
-                Text = "Save",
-            };
-            save.Clicked += Save_Clicked;
-            yield return save;
-        }
-
         private async void Save_Clicked(object? sender, EventArgs e)
         {
             if (context == null)
                 return;
             if (!context.ChangeTracker.HasChanges())
-                await DisplayAlert($"No changes were made.", "", "Ok");
-            await context.SaveChangesAsync();
+                await DisplayAlert($"No changes were made.", "", "Ok");//TODO dont wait if dont care about result
+            await context.SaveChangesAsync();//TODO dont save if no changes
             await Navigation.PopAsync();
         }
